move sorted tests
